Move WildcardTypeName
